feat: 예약 취소 및 환불 기능

+++ src/main/java/org/example/oristationbackend/dto/user/PayCancelDto.java
@@ -12,5 +12,5 @@ public class PayCancelDto {
     String imp_uid;
     String merchant_uid;
     int amount;
-    Long checksum;
+    int checksum;
 }
